# openAPI.yaml
openapi: 3.0.3
info:
  title: Bende REST interface
  version: 1.0.0
servers:
  - url: http://localhost:8081
paths:
  /konnektors/{konnektorId}:
    parameters:
      - name: konnektorId
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getKonnektor
      responses:
        200:
          description: konnektor retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KonnektorDTO"
        404:
          description: no konnektor found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    put:
      operationId:  updateKonnektor
      summary: Update the complete konnektor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KonnektorDTO'
      responses:
        204:
          description: Konnektors updated successfully
        403:
          description: Forbidden
        404:
          description: Konnektor not found
        500:
          description: Server Error
    patch:
      operationId: updateKonnektorHostname
      summary: updates hostname only
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KonnektorHostnameDTO'
      responses:
        204:
          description: Konnektors updated successfully
        403:
          description: Forbidden
        404:
          description: Konnektor not found
        500:
          description: Server Error
    delete:
      description: deleteing a konnektor
      operationId: deleteKonnektor
      responses:
        '200':
          description: Konnektor deleted
  /konnektors:
    get:
      description:  gets all konnektors
      operationId: getAllKonnektors
      parameters:
        - in: query
          name: hostName
          schema:
            type: string
          description: hostName of the connector
        - in: query
          name: serialNumber
          schema:
            type: string
          description: serialNumber of the connector
        - in: query
          name: firmwareVersion
          schema:
            type: string
          description: firmwareVersion of the connector
        - in: query
          name: hardwareVersion
          schema:
            type: string
          description: hardwareVersion of the connector
        - in: query
          name: created
          schema:
            type: string
            format: date-time
          description: created of the connector
      responses:
        200:
          description: all konnektors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KonnektorDTO"
        404:
          description: not konnektors found
    post:
      description: creates a new konnektor
      operationId:  createKonnektor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KonnektorDTO"
      responses:
        204:
          description: konnektor created
        500:
          description: server error
  /auditlogs:
    post:
      operationId: createAuditLog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref:  "#/components/schemas/CreateAuditLogRequestDTO"
      responses:
        204:
          description: AuditLog created
        400:
          description: Konnektor not found
        500:
          description: Server error
  /auditlogs/{auditlogId}:
    parameters:
      - name: auditlogId
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getAuditLogs
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLogDTO"
    put:
      operationId: updateAuditlog
      summary: Update the complete auditlog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuditLogDTO'
      responses:
        204:
          description: auditlog updated successfully
        403:
          description: Forbidden
        404:
          description: auditlog not found
        500:
          description: Server Error
components:
  parameters:
    konnektorId:
      name: konnektorId
      in: path
      schema:
        type: integer
        # format: int64 # should be "long"?
      required: true
  schemas:
    ApiError:
      type: object
      properties:
        status:
          type: string
        timestamp:
          type: string
        message:
          type: string
    ArtistResponseDTO:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
      required:
        - id
        - username
        - role
    EmployeesResponseDTO:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstname:
          type: string
        lastname:
          type: string
      required:
        - id
    AuditLogDTO:
      type: object
      properties:
        id:
          type: integer
        user:
          type: string
        konnektor:
          type: integer
        userAction:
          $ref:  "#/components/schemas/AuditLogMessageDTO"
        timestamp:
          type: string
          format: date-time
    UpdateKonnektorRequestDTO:
      type: object
      properties:
        konnektorId:
          type: integer
          format: int64
        hostName:
          type: string
        serialNumber:
          type: string
        firmwareVersion:
          type: string
        hardwareVersion:
          type: string
        active:
          type: boolean
        created:
          type: string
          format: date-time
      required:
        - konnektorId
    CreateAuditLogRequestDTO:
      type: object
      properties:
        user:
          type: string
        userAction:
          $ref:  "#/components/schemas/AuditLogMessageDTO"
        konnektor:
          type: integer
      required:
        - user
        - userAction
    KonnektorDTO:
      type: object
      properties:
        id:
          type: integer
        hostName:
          type: string
        serialNumber:
          type: string
        firmwareVersion:
          type: string
        hardwareVersion:
          type: string
        active:
          type: boolean
        created:
          type: string
          format: date-time
        validUntil:
          type: string
          format: date-time
        auditlogs:
          type: array
          items:
            $ref: "#/components/schemas/AuditLogDTO"
      required:
        - hostName
    KonnektorHostnameDTO:
      type: object
      properties:
        hostName:
          type: string
      required:
        - hostName
    ErrorDTO:
      type: object
      properties:
        message:
          type: string
    AuditLogMessageDTO:
      type: string
      enum:
        - ADD_NEW_KONNEKTOR
        - REMOVE_KONNEKTOR
        - UPDATE_KONNEKTOR_SETTINGS
        - USER_LOGIN
        - USER_LOGOUT
        - CREATE_USER
        - UPDATE_USER_DATA
        - DELETE_USER